Fix everything and make it run
