+++ packages/auth/src/auth-options.ts
@@ -50,6 +50,9 @@ export const getAuthOptions = ({
     async session({ session, token }) {
       if (token.sub) {
         const user = await core.userService.getUserBySubject(token.sub)
+        if (user === undefined) {
+          throw new Error(`Found no matching user for ${token.sub}`)
+        }
         session.user.id = user.id
         session.sub = token.sub
       }
