+++ packages/core/vitest-integration.setup.ts
@@ -2,19 +2,8 @@ import { beforeEach } from "vitest"
 import { PostgreSqlContainer, type StartedPostgreSqlContainer } from "@testcontainers/postgresql"
 import { createMigrator, createKysely } from "@dotkomonline/db"
 import { createEnvironment } from "@dotkomonline/env"
-import { type MigrationResultSet } from "kysely"
 
 const containers: StartedPostgreSqlContainer[] = []
-const handleResult = (result: MigrationResultSet) => {
-  const executedMigrations = result.results ?? []
-  for (const migration of executedMigrations) {
-    console.info(`${migration.direction} - Migrated ${migration.migrationName} - ${migration.status}`)
-  }
-  if (result.error) {
-    console.warn("Caught error during migration run:")
-    console.warn(JSON.stringify(result.error, null, 2))
-  }
-}
 
 beforeEach(async () => {
   const container = await new PostgreSqlContainer("public.ecr.aws/z5h0l8j6/dotkom/pgx-ulid:0.1.3")
@@ -26,12 +15,11 @@ beforeEach(async () => {
   process.env.DATABASE_URL = container.getConnectionUri()
   const env = createEnvironment()
   const kysely = createKysely(env)
-  const migrator = createMigrator(kysely)
-  const result = await migrator.migrateToLatest()
-  handleResult(result)
+  const migrator = createMigrator(kysely, new URL("node_modules/@dotkomonline/db/src/migrations", import.meta.url))
+  await migrator.migrateToLatest()
   containers.push(container)
 })
 
 process.on("exit", async () => {
-  await Promise.all(containers.map((container) => container.stop()))
+  await Promise.all(containers.map(async (container) => container.stop()))
 })
