+++ packages/db/src/migrator.ts
@@ -1,14 +1,14 @@
-import { promises as fs } from "fs"
-import * as path from "path"
+import fs from "fs/promises"
+import path from "path"
 import { FileMigrationProvider, type Kysely, Migrator } from "kysely"
 import { type Database } from "./"
 
-export const createMigrator = (db: Kysely<Database>) =>
+export const createMigrator = (db: Kysely<Database>, url = new URL("migrations", import.meta.url)) =>
   new Migrator({
     db,
     provider: new FileMigrationProvider({
       fs,
       path,
-      migrationFolder: new URL("migrations", import.meta.url).pathname,
+      migrationFolder: url.pathname,
     }),
   })
