+++ packages/db/src/utils.d.ts
@@ -1,0 +1,13 @@
+import { type CreateTableBuilder, type SchemaModule } from "kysely"
+
+interface DefaultOptions {
+  id?: boolean
+  createdAt?: boolean
+  updatedAt?: boolean
+}
+
+export const createTableWithDefaults: (
+  tableName: string,
+  options: DefaultOptions,
+  schema: SchemaModule
+) => CreateTableBuilder<string>
